remove tool for repulsion. fix bug in delete tool
